Update StudentController.java
Added GetMapping to allow the search of a student based on their student_id field. Still need to add capability to add a student if they are not found during the search based on id. Also, I need to change the search criteria to search by email address not student_id